support writing integer64 values, closes #243
- Values of class `"integer64"` are now supported for `dbWriteTable()` and `dbBind()` (#243).

+++ src/integer64.h
@@ -3,7 +3,7 @@
 
 #define INT64SXP REALSXP
 
-#define NA_INTEGER64 (0x8000000000000000)
+#define NA_INTEGER64 (int64_t(0x8000000000000000))
 
 inline int64_t* INTEGER64(SEXP x) {
   return reinterpret_cast<int64_t*>(REAL(x));
